+++ src/components/QuranReader/index.tsx
@@ -13,20 +13,17 @@ import classNames from 'classnames';
 import { selectTafsirs, TafsirsSettings } from 'src/redux/slices/QuranReader/tafsirs';
 import { getDefaultWordFields } from 'src/utils/api';
 import { selectIsUsingDefaultReciter, selectReciter } from 'src/redux/slices/AudioPlayer/state';
-import Verse from 'types/Verse';
-import clipboardCopy from 'clipboard-copy';
 import { makeJuzVersesUrl, makePageVersesUrl, makeVersesUrl } from 'src/utils/apiPaths';
 import { buildQCFFontFace, isQCFFont } from 'src/utils/fontFaceHelper';
 import { QuranReaderStyles, selectQuranReaderStyles } from 'src/redux/slices/QuranReader/styles';
 import { selectReadingPreference } from 'src/redux/slices/QuranReader/readingPreferences';
-import { getWordDataByLocation } from 'src/utils/verse';
 import ReadingView from './ReadingView';
 import TranslationView from './TranslationView';
 import { QuranReaderDataType, ReadingPreference } from './types';
 import Notes from './Notes/Notes';
 import styles from './QuranReader.module.scss';
 import TafsirView from './TafsirView';
-import { DATA_ATTRIBUTE_WORD_LOCATION } from '../dls/QuranWord/QuranWord';
+import onCopyQuranWords from './onCopyQuranWords';
 // import ContextMenu from './ContextMenu';
 
 type QuranReaderProps = {
@@ -119,40 +116,11 @@ const QuranReader = ({
     }
   };
 
-  /**
-   * select all DOM node that contains data attribute `data-word-location`
-   * then see if that node is within `selection`
-   * then get the uthmani_text for those nodes
-   * then copy the uthmani_text to clipboard
-   *
-   * We need to use `uthmani_text` instead of the rendered text
-   * because sometime user is using code_v1 and code_v2 font,
-   * which uses unicode, so it can't be copied properly by user
-   */
-  const onCopy = (event) => {
-    const selection = document.getSelection();
-    const quranWordsToCopy = Array.from(
-      document.querySelectorAll(`[${DATA_ATTRIBUTE_WORD_LOCATION}]`),
-    )
-      .filter((node) => selection.containsNode(node, true))
-      .map((node) => {
-        const wordLocation = node.getAttribute(DATA_ATTRIBUTE_WORD_LOCATION);
-        return getUthmaniText(wordLocation, verses);
-      });
-
-    if (quranWordsToCopy.length > 0) {
-      // only do prevent default, call clipboardCopy when there are quran words to copy
-      // otherwise, user can't copy the translation text
-      event.preventDefault();
-      clipboardCopy(quranWordsToCopy.join(' '));
-    }
-  };
-
   return (
     <>
       {/* <ContextMenu /> */}
       <div
-        onCopy={onCopy}
+        onCopy={(event) => onCopyQuranWords(event, verses)}
         className={classNames(styles.container, { [styles.withVisibleSideBar]: isSideBarVisible })}
       >
         <div className={styles.infiniteScroll}>
@@ -250,19 +218,3 @@ const getRequestKey = ({
 };
 
 export default QuranReader;
-
-const getUthmaniText = (wordLocation: string, verses: Verse[]) => {
-  const [chapter, verse, location] = getWordDataByLocation(wordLocation);
-
-  // find by verseKey
-  const selectedVerse = verses.find((v) => v.verseKey === `${chapter}:${verse}`);
-  if (!selectedVerse) return '';
-
-  // find by location
-  const selectedWord = selectedVerse.words.find(
-    (word) => Number(word.position) === Number(location),
-  );
-  if (!selectedWord) return '';
-
-  return selectedWord.textUthmani;
-};
